Feature_Pipeline/pyproject.toml
[tool.poetry]
name = "feature-pipeline"
version = "0.1.0"
description = ""
authors = ["pavanmodi <pmodi1203@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
hopsworks = "^3.7.0"
yarl = "^1.17.2"
fire = "^0.7.0"
requests = "^2.32.3"
pandas = ">=1.3.5"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
